kernel: sched: Align 'tdq_lock' on cache line size
The lock should be aligned on a cache line boundary so it isn't in a
cache line with other data. This is important for 'tdq_lock' as it is
used by every processor which constantly acquires and releases it. This
alignment can reduce how violently cache lines bounce between processor
local caches.

Signed-off-by: Ian Moffett <ian@osmora.org>