kernel/amd64: Add `alloc` to pmap_update_tbl()
Add an `alloc' paramater to pmap_update_tbl() to ensure that functions
like pmap_map() and pmap_unmap() have control over, whether or not to
allocate new page table entries. This is useful as unmapping memory does
not require new page table entries to be allocated and doing so anyways
may result in undefined behaviour such as hangs.

Signed-off-by: Ian Moffett <ian@osmora.org>