kernel: exit: Only kill leaf if not exiting
During an exit(), the parent needs to kill all of the child processes
that depend on it. However, there might be an occurrence where one or
more is already exiting and trying to kill such processes would be
problematic and may result in issues like double frees. This commit
ensures that the exiting parent only kills child processes that aren't
already exiting themselves.

Signed-off-by: Ian Moffett <ian@osmora.org>