usr.bin: osh: Implement shell script parsing
Previously, OSH was only capable of interpreting commands that came
directly from the console. This commit allows calling /usr/bin/osh while
passing a path to a script to be parsed and interpreted. This commit
additionally adds a `command_match()` routine to simplify the matching of
commands. On top of this, cmd_run() now also takes a `wait` argument
that allows the caller to specify whether not to daemonize the spawned
process or wait on its execution to finish.

Signed-off-by: Ian Moffett <ian@osmora.org>