kernel: Introduce SYS_inject syscall
The SYS_inject syscall will allow the user to inject Kernel Runtime Quantums
(KRQs) into the running kernel. As of now, set paths are not supported and the
initial implementation includes a NULL path used to load all deferrable drivers.

Previously during kernel startup, all deferrable drivers were loaded
automatically. This commit makes it so that they are only loaded when
sys_inject() is called with a NULL path.

Signed-off-by: Ian Moffett <ian@osmora.org>