kernel: ahci: Lower link wait time to 8 ms
Section 3.3.11 of the AHCI spec states that software must leave the DET
field set for a minimum of 1 milisecond. Most hardware conforms to this.
Previously, our buffer time was way too high and would slow down the
time it takes for the AHCI driver to finish initialization. If any SATA
drive takes longer than this to start up, fuck it, something is wrong
with it.

Signed-off-by: Ian Moffett <ian@osmora.org>