kernel/amd64: mp: Pin one idle thread to each core
Each core on the CPU gets one idle thread so that it is always up to
something. To ensure these threads are not bouncing wildly between cores
and are always ready for their respective processor to take when they
have no threads to grab. We must call proc_pin() on each for every
processor.

Signed-off-by: Ian Moffett <ian@osmora.org>