kernel: Introduce initial VSR impl
This commit introduces the initial implementation of Virtual System
Resources (VSRs). Virtual system resources enable specific subsystems to
control global state while the VSR acts as a middle man between the
process, and the actual resource itself. This allows the products of
certain operations to be routed to either the global state or a local
(per-process) "shallow copy" whose data and side effects are only
visible through that process. An example of something this would be
useful for is some sort of transparent access control mechanism where
if a process should not modify global system-wide state, it can at least
modify the state for itself.

Signed-off-by: Ian Moffett <ian@osmora.org>