kernel/amd64: Limit frame depth in backtrace
A backtrace usually occurs during critcal system events like panics.
When the system in a error state, we cannot trust the correctness of the
stack. This commit aims to improve checks done in a731d4e by adding a
max frame depth that limits how many frames can be walked.

Signed-off-by: Ian Moffett <ian@osmora.org>