usr: libc: Add initial disk engine interface
This commit implements a libc interface to the Hyra disk engine and
adds functions such as disk_read() and disk_write(). There is also a
mostly internal __disk_io() routine which is made to have external
linkage as it could potentially be useful for cases where fine grained
control over the engine is needed, though it is recommended to rely on
the other functions whenever possible.

Signed-off-by: Ian Moffett <ian@osmora.org>