kernel: disk: Use 'void *' in disk_param_init()
The disk parameters allow data to be passed into the disk engine via its
interface. The data may be raw byte streams from the drive, or bytes to
be written to the drive. However not all the time is the data to be an
arbitrary stream of bytes and may contain structured data instead.

Signed-off-by: Ian Moffett <ian@osmora.org>